+++ python/pyarrow/includes/libarrow_s3fs.pxd
